lint: rename unused bindings and simplify idioms
Behaviour-preserving ruff fixes, previewed before applying:
- RUF059/B007: unused unpacked and loop variables get a leading `_`.
- RUF007: zip(x, x[1:]) in tests is itertools.pairwise.
- SIM108/SIM103/SIM110/SIM105: small if/else and loop bodies become a
  conditional expression, a returned condition, any(), or
  contextlib.suppress.
- RUF015: tests take the first match with next() instead of [...][0].
- RUF046: round() already returns an int, so int(round(x)) is round(x).

+++ config/config.py
@@ -2776,7 +2776,7 @@ def setfontsizes(config: CalendarConfig) -> CalendarConfig:
     )
 
     # Watermark base font size (paper-size aware, theme-overridable)
-    config.watermark_font_size = int(round(_clamp(h * 0.10, 24.0, 256.0)))
+    config.watermark_font_size = round(_clamp(h * 0.10, 24.0, 256.0))
 
     # Phase 2 wave 2: inject heuristic-derived size tokens so renderers'
     # `tk.get("size")` reads always return a value, even from test fixtures
